requirements.txt
matplotlib==3.3.3
numpy==1.18.5
Pillow==8.0.1
tensorflow==2.3.0
pydot==1.4.1
